Update Log:
-Começando a fazer os métodos de calcular;

+++ src/Model/Calculadora.java
@@ -1,25 +1,25 @@
 package Model;
 
 public class Calculadora {
-    private Float valor1;
-    private Float valor2;
+    private Double valor1;
+    private Double valor2;
 
     public Calculadora() {
     }
 
-    public Float getValor1() {
+    public Double getValor1() {
         return valor1;
     }
 
-    public Float getValor2() {
+    public Double getValor2() {
         return valor2;
     }
 
-    public void setValor1(Float valor1) {
+    public void setValor1(Double valor1) {
         this.valor1 = valor1;
     }
 
-    public void setValor2(Float valor2) {
+    public void setValor2(Double valor2) {
         this.valor2 = valor2;
     }
